+++ src/main/kotlin/rayTraceTypescript/Interval.kt
@@ -1,9 +1,9 @@
 package rayTraceTypescript
 
-class Interval(val min: Double, val max: Double) {
-    fun contains(x: Double): Boolean = x in min..max
-    fun surrounds(x: Double): Boolean = x > min && x < max
-    fun clamp(x: Double): Double = when {
+class Interval(val min: Float, val max: Float) {
+    fun contains(x: Float): Boolean = x in min..max
+    fun surrounds(x: Float): Boolean = x > min && x < max
+    fun clamp(x: Float): Float = when {
         x < min -> min
         x > max -> max
         else -> x
