+++ src/main/kotlin/rayTraceTypescript/objects/Hit.kt
@@ -5,7 +5,7 @@ import rayTraceTypescript.Ray
 import rayTraceTypescript.Vector
 import rayTraceTypescript.materials.Material
 
-class Hit(ray: Ray, val point: Point, outwardNormal: Vector, val material: Material, val t: Double) {
+class Hit(ray: Ray, val point: Point, outwardNormal: Vector, val material: Material, val t: Float) {
     var normal: Vector = outwardNormal
     var frontFace: Boolean = true
 
@@ -14,7 +14,7 @@ class Hit(ray: Ray, val point: Point, outwardNormal: Vector, val material: Mater
     }
 
     fun setFrontFace(ray: Ray, outwardNormal: Vector) {
-        frontFace = Vector.dotProduct(ray.direction, outwardNormal) < 0.0
+        frontFace = Vector.dotProduct(ray.direction, outwardNormal) < 0.0f
         normal = if (frontFace) outwardNormal else outwardNormal.negate()
     }
 }
